Add note about additional failure mode
On a Pixel 4XL I was successfully scheduling notifications but they were
never showing. Turns out this was because I was missing some of the
release build setup. The notification was firing but not showing,
because the system errored out when it couldn't find the resource.

This commit adds a note to the README about this failure mode.