[flutter_local_notifications] Fix macOS deprecation warnings (#1965)
* Add some MARKs

Helps to navigate in Xcode when you can see groupings of code.

* Mark macOS NSUserNotification methods unavailable (#1858)

NSUserNotification was deprecated in macOS 10.14. Mark any OS X target methods
using this API as deprecated. This eliminates compile-time warnings when
building with 10.14 or higher.

* Fix deprecation warning statements

Specified the wrong version! Fixed, and added the introduction version as well.
Also switching from OSX to macOS. All Swift documentation uses macOS and not
OSX. Figured best to follow the official docs, even though OSX is valid.

* Use macOS versus OSX for Swift availability

Following the previous commit, standardize on using "macOS" for availability in
Swift code. This matches the official Swift language documentation, versus
"OSX" which, while valid, is present for backward compatibility, but not
mentioned in the Swift docs.