Attempt to add providesAppNotificationSettings support for Macos Example app (doesn't work)
This commit is a proof-of-concept in which I attempted to implement and demonstrate whether configuring this "Configure in App" API on macOS has any actual effect (like it does on iOS): I can confirm, it does *not*.

Details:
Even with these modifications and corresponding macOS AppDelegate configuration, no such "Configure in App" option was visible upon right-clicking the notification in the macOS notification pane, as tested on macOS Tahoe 26.0.

Additionally, I could not find any examples on the web demonstrating this feature on macOS.

As such, while macOS does support the providesAppNotificationSettings API (likely for API symmetry with iOS, and for potential future compatibility), at the moment, this API does not result in any user-facing UI/UX change on macOS.

This commit will be reverted, but I'll preserve it in the branch as a record of the finding.